+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/signaling/message/Catalog.kt
@@ -44,6 +44,7 @@ data class ConnectMessage(
     val dataChannelSignaling: Boolean? = null,
     @SerializedName("ignore_disconnect_websocket")
     val ignoreDisconnectWebsocket: Boolean? = null,
+    @SerializedName("data_channels") val dataChannels: List<Map<String, Any>>? = null,
     @SerializedName("redirect") var redirect: Boolean? = null
 )
 
@@ -162,10 +163,6 @@ data class NotificationMessage(
     @SerializedName("metadata_list") val metadataList: Any?,
     @SerializedName("minutes") val connectionTime: Long?,
     @SerializedName("channel_connections") val numberOfConnections: Int?,
-    @Deprecated("numberOfUpstreamConnections は 2021 年 6 月リリース予定の Sora にて廃止されます。")
-    @SerializedName("channel_upstream_connections") val numberOfUpstreamConnections: Int?,
-    @Deprecated("numberOfDownstreamConnections は 2021 年 6 月リリース予定の Sora にて廃止されます。")
-    @SerializedName("channel_downstream_connections") val numberOfDownstreamConnections: Int?,
     @SerializedName("channel_sendrecv_connections") val numberOfSendrecvConnections: Int?,
     @SerializedName("channel_sendonly_connections") val numberOfSendonlyConnections: Int?,
     @SerializedName("channel_recvonly_connections") val numberOfRecvonlyConnections: Int?,